Handled internal objects in the volume.
Render the volume last. Pass depth and framebuffer to render for blending
The volume is set as transparent to make VTK render it after
everything else. Current framebuffer and z-buffer is copied to
textures and passed to the shader so existing objects can be correctly
blended with the volume

If the ray hits the depth map the background pixel at that value is
blended in as a opaqe voxel and the ray is terminated. This allows us
to remove the previous trick with gl_FragDepth which disabled the GPUs
early z test and allows us to drop our buggy version of an early z
test. Result: Better images and the same performance


Former-commit-id: e68ba6f8c05de59348b665152b548e4d5d1186ec